feat: update the 'updateMovieHandler' so that it supports partial updates of movie records
The updateMovieHandler had to get the complete JSON request with all the
parameter of the movie (title,year,...) for it to be able to update the
record.
Now, the client can just send the parameter that they want to change and
other values will be filled automatically.

The handler method is also changed from PUT to the appropriate PATCH
method.