fix: possible race condition when updating a record
If two users update a movie record at the same time, then a data race
condition is introduced.
This is fixed by checking the last updated version when updating a
resource. If the version is as expected, then the record is updated,
else an edit conflict error is sent to the client.